# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  CreatePostRequest:
    properties:
      content:
        type: string
      image_upload_ids:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        maxLength: 150
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  ErrorCode:
    enum:
    - PARSE_ERROR
    - VALIDATION_ERROR
    - CONFLICT_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - UNPROCESSABLE_ENTITY
    - RATE_LIMITED
    - INTERNAL_SERVER_ERROR
    type: string
    x-enum-varnames:
    - ParseCode
    - ValidationCode
    - ConflictCode
    - UnauthorizedCode
    - ForbiddenCode
    - NotFoundCode
    - InvalidCode
    - RateLimitedCode
    - InternalServerCode
  ErrorInfo-any:
    properties:
      code:
        $ref: '#/definitions/ErrorCode'
      details: {}
      message:
        type: string
    type: object
...
  ErrorResponse-any:
    properties:
      error:
        $ref: '#/definitions/ErrorInfo-any'
    type: object
...
  PostImageResponse:
    properties:
      position:
        type: integer
      upload_id:
        type: string
      url:
        type: string
    type: object
  PostResponse:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/PostImageResponse'
        type: array
      liked_by_me:
        type: boolean
      likes_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  Response-array_PostResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/PostResponse'
        type: array
    type: object
...
  Response-internal_post_PostResponse:
    properties:
      data:
        $ref: '#/definitions/PostResponse'
    type: object
...
  UpdatePostRequest:
    properties:
      content:
        type: string
      image_upload_ids:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        maxLength: 150
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Social media API for posts, authentication, and feed management
  title: Artblog API
  version: "1.0"
paths:
...
  /posts:
    post:
      consumes:
      - application/json
      description: Creates a new post with optional carousel images (max 10).
      parameters:
      - description: Create post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Response-internal_post_PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Deletes a post and marks associated media as superseded.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      description: Retrieves post details including author and ordered carousel images.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response-internal_post_PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      summary: Get post details
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Updates post title, content, and/or reconciles the image carousel.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Update post request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response-internal_post_PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      security:
      - BearerAuth: []
      summary: Update post
      tags:
      - posts
...
  /posts/{id}/like:
    delete:
      description: Removes the authenticated user's like from a post. Repeating the
        request is safe.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      security:
      - BearerAuth: []
      summary: Unlike post
      tags:
      - posts
    put:
      description: Adds the authenticated user's like to a post. Repeating the request
        is safe.
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      security:
      - BearerAuth: []
      summary: Like post
      tags:
      - posts
  /posts/author/{author_id}:
    get:
      description: Returns a paginated list of posts by a specific author.
      parameters:
      - description: Author UUID
        in: path
        name: author_id
        required: true
        type: string
      - description: Number of posts to return (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response-array_PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      summary: List posts by author
      tags:
      - posts
  /posts/recent:
    get:
      description: Returns a paginated list of recent posts with images.
      parameters:
      - description: Number of posts to return (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response-array_PostResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse-any'
      summary: List recent posts
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"